requirements.txt
asyncio==4.0.0
aiogram==3.25.0
requests==2.32.5
python-dotenv==1.2.1
asyncpg==0.31.0
apscheduler==3.11.2